fix: tree shaking performance, package support - improve icon tree shaking - reduce duplicate core component styles - clean package output for better support

+++ packages.js
@@ -14,8 +14,11 @@ const read = (dir) => fs.readdirSync(dir).reduce((files, file) =>
     files.concat(path.join(dir, file)),
   []);
 
-fs.copySync('./dist/lithium/esm2015', './dist/lithium');
-['bundles','esm5', 'fesm5', 'esm2015', 'fesm2015'].forEach(f => fs.removeSync(`./dist/lithium/${f}`));
+if (fs.existsSync('./dist/lithium/esm2015')) {
+  fs.copySync('./dist/lithium/esm2015', './dist/lithium');
+  ['bundles','esm5', 'fesm5', 'esm2015', 'fesm2015'].forEach(f => fs.removeSync(`./dist/lithium/${f}`));
+}
+
 read('./dist/lithium').filter(f => f.includes('lithium-ui') || f.includes('public-api')).forEach(f => fs.removeSync(f));
 read('./dist/lithium').filter(f => f.includes('package.json')).forEach(file => {
   const data = fs.readJsonSync(file);
@@ -23,7 +26,6 @@ read('./dist/lithium').filter(f => f.includes('package.json')).forEach(file => {
   data.module = './index.js';
   data.main = './index.js';
   data.type = 'module';
-  data.sideEffects = data.name.includes('/') ? undefined : data.sideEffects;
   ['__processed_by_ivy_ngcc__', 'scripts', 'es2015', 'esm5', 'esm2015', 'fesm5', 'fesm2015'].forEach(p => delete data[p]);
   fs.writeJsonSync(file, data, { spaces: 2 });
 });
